refactor, add custom merger
