+++ data/aave-oracles.json
@@ -1,0 +1,4 @@
+{
+  "names": {},
+  "shortNames": {}
+}
