+++ data/aave-tokens.json
@@ -1,0 +1,4 @@
+{
+  "names": {},
+  "shortNames": {}
+}
