+++ data/update-manifest.json
@@ -1,32 +1,28 @@
 {
-  "version": "2025-08-23T12-32-51-317Z",
-  "generatedAt": "2025-08-23T12:32:51.318Z",
-  "totalAdded": 735,
-  "totalUpdated": 0,
+  "version": "2025-08-23T13-37-04-589Z",
+  "generatedAt": "2025-08-23T13:37:04.589Z",
+  "totalAdded": null,
+  "totalUpdated": null,
   "updatedFiles": [
-    "./lender-labels.json",
-    "./morpho-oracles.json",
-    "./aave-labels.json",
-    "./aave-oracles.json"
+    "./data/lender-labels.json",
+    "./data/morpho-oracles.json"
   ],
   "results": [
     {
-      "targetFile": "./lender-labels.json",
+      "targetFile": "./data/lender-labels.json",
       "added": 730,
       "updated": 0
     },
     {
-      "targetFile": "./morpho-oracles.json",
-      "added": 5,
-      "updated": 0
+      "targetFile": "./data/morpho-oracles.json"
     },
     {
-      "targetFile": "./aave-labels.json",
+      "targetFile": "./data/aave-tokens.json",
       "added": 0,
       "updated": 0
     },
     {
-      "targetFile": "./aave-oracles.json",
+      "targetFile": "./data/aave-oracles.json",
       "added": 0,
       "updated": 0
     }
